LinkInfo class in new file. Add Unit Test for LinkInfo class.
