+++ UnitTests/CMSWebTests/CMSWebTests.csproj
@@ -122,6 +122,7 @@
     <Compile Include="Areas\Main\Controllers\EmailControllerTests.cs" />
     <Compile Include="Areas\Manage\Models\Involvement\InvolvementTabModelTests.cs" />
     <Compile Include="Areas\Manage\Models\AccountModelTests.cs" />
+    <Compile Include="Areas\OnlineReg\Models\LinkInfoTests.cs" />
     <Compile Include="Areas\OnlineReg\Models\OnlineRegModelTest.cs" />
     <Compile Include="Areas\OnlineReg\Models\OnlineRegPerson\HelperTests.cs" />
     <Compile Include="Areas\OnlineReg\Models\OnlineRegPerson\OnlineRegPersonModelTests.cs" />
